Move the MR class loading logic to MapReduceClassLoader to avoid name conflict with the ApplicationClassLoader, which the ordering might get messed up in unit-test.
